style: concatenate routes modules
